feat: 상세 필터 컴포넌트 분리
