+++ src/components/listPage/FilterDropdown.module.scss
@@ -1,0 +1,151 @@
+.filterContainer {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 24px 20px;
+  gap: 10px;
+  z-index: 1;
+  position: absolute;
+  top: 130%;
+  right: 0;
+  width: 390px;
+  background: #ffffff;
+  border: 1px solid #e5e4e7;
+  box-shadow: 0px 2px 8px rgba(120, 116, 134, 0.25);
+  border-radius: 10px;
+
+  font-family: 'Spoqa Han Sans Neo';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 25px;
+  color: #111322;
+}
+.fliterTop {
+  position: relative;
+  display: flex;
+  width: 350px;
+  margin-bottom: 14px;
+  gap: 10px;
+  align-items: center;
+}
+.xImage {
+  margin: 0 0 0 auto;
+}
+.filterName {
+  font-family: 'Spoqa Han Sans Neo';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 26px;
+  color: #111322;
+}
+.filterLocation {
+  width: 350px;
+  height: 258px;
+  background: #ffffff;
+  border: 1px solid #e5e4e7;
+  border-radius: 6px;
+  overflow: auto;
+}
+.line {
+  width: 350px;
+  height: 2px;
+  background: #f2f2f3;
+  margin: 10px 0 10px;
+}
+.filterDate {
+  padding: 10px;
+  font-size: 16px;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  width: 350px;
+
+  &:focus {
+    outline: none;
+    border-color: var(--yellow5);
+  }
+
+  &::placeholder {
+    color: #999;
+  }
+}
+.filterPrice {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  padding: 16px 20px;
+  gap: 10px;
+  width: 169px;
+  height: 58px;
+  background: #ffffff;
+  border: 1px solid #cbc9cf;
+  border-radius: 6px;
+
+  font-family: 'Spoqa Han Sans Neo';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 26px;
+  color: #111322;
+
+  &:focus {
+    border: 1px solid var(--yellow5);
+    outline: none;
+  }
+  &::placeholder {
+    font-size: 16px;
+    line-height: 26px;
+  }
+}
+.filterPriceText {
+  font-family: 'Spoqa Han Sans Neo';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 26px;
+  color: #111322;
+  position: absolute;
+  left: 140px;
+}
+.filterBottom {
+  display: flex;
+  gap: 10px;
+}
+.reset {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+  width: 82px;
+  height: 48px;
+  background: #ffffff;
+  border: 1px solid var(--yellow5);
+  border-radius: 6px;
+
+  font-family: 'Spoqa Han Sans Neo';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 16px;
+  line-height: 20px;
+  color: var(--yellow5);
+}
+.submit {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+  width: 260px;
+  height: 48px;
+  background: var(--yellow5);
+  border-radius: 6px;
+  border: none;
+
+  font-family: 'Spoqa Han Sans Neo';
+  font-style: normal;
+  font-weight: 700;
+  font-size: 16px;
+  line-height: 20px;
+  color: #ffffff;
+}
